refactor: centralize notification loading without changing modes (#920)
